requirements.txt
gradio>=4.44.0
transformers>=4.41.0
torch
sentence-transformers>=2.7.0
faiss-cpu
pypdf
numpy